+++ tests/blocks/FileListBlockTest.php
@@ -28,6 +28,15 @@ class FileListBlockTest extends GenericBlockTestCase
         $this->assertStringContainsString('<ul><li><a target="_blank" href="path/to/image.jpg">foobar</a></li></ul>', $this->renderFrontendNoSpace());
     }
 
+    public function testOrderedList()
+    {
+        $this->block->setVarValues(['listType' => 'ol']);
+        $this->block->addExtraVar('fileList', [
+            ['href' => 'path/to/image.jpg', 'caption' => 'foobar', 'extension' => 'jpg'],
+        ]);
+        $this->assertStringContainsString('<ol><li><a target="_blank" href="path/to/image.jpg">foobar</a></li></ol>', $this->renderFrontendNoSpace());
+    }
+
     public function testFilesWithSuffix()
     {
         $this->block->setCfgValues(['showType' => 1]);
@@ -35,5 +44,14 @@ class FileListBlockTest extends GenericBlockTestCase
                 ['href' => 'path/to/image.jpg', 'caption' => 'foobar', 'extension' => 'jpg'],
         ]);
         $this->assertStringContainsString('<ul><li><a target="_blank" href="path/to/image.jpg">foobar (jpg)</a></li></ul>', $this->renderFrontendNoSpace());
+    }
+
+    public function testListWithCssClasses()
+    {
+        $this->block->setCfgValues(['listCssClass' => 'foo', 'elementCssClass' => 'bar']);
+        $this->block->addExtraVar('fileList', [
+            ['href' => 'path/to/image.jpg', 'caption' => 'foobar', 'extension' => 'jpg'],
+        ]);
+        $this->assertStringContainsString('<ul class="foo"><li class="bar"><a target="_blank" href="path/to/image.jpg">foobar</a></li></ul>', $this->renderFrontendNoSpace());
     }
 }